Reimplement memory map algorithm with functional decomposition
Go back and redo the entire algorithm for turning the `chosen`
properties into the linker script parameterization using functional
decomposition.

I wrote down all of the steps in between a Devicetree and the resulting
data structures for the template rendering in theory_of_operation.md.
Then I created a separate module memory_map.py which functionally
decomposes that algorithm and implements it step-by-step.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>