Add support for picolibc's TLS data
Picolibc uses the TLS support from the toolchain for thread local
storage, so it needs the first TLS data block allocated and
initialized during application startup.

This block is carefully placed between initialized data and
uninitialized data so that the existing data and bss initialization
sequences will also set up the TLS data. All the application needs to
do is to assign the TP register to the value of the __tls_base symbol.

This required adding a PHDRS section to the linker script so that the
TLS sections could get a PT_TLS attribute associated with them.

Signed-off-by: Keith Packard <keithp@keithp.com>