ALIGN(16) instead of ALIGN(8)
Why do we keep picking `ALIGN(8)`?

It is my reading of the (System V) ABI that the alignment of a section shall be at least as strict as its most-strictly aligned member. The strictest alignment defined by the RISC-V psABI is `16 bytes == alignof(max_align_t)`: 
https://github.com/riscv/riscv-elf-psabi-doc/blob/master/riscv-elf.md#-c-type-sizes-and-alignments 

My concern is that a compiler can (correctly) emit code assuming 16-byte alignment of some object, and this code is subsequently broken by a link script generated from this Jinja template. (@jim-wilson: do you think this possible?) Unfortunately I'm not clever enough to construct a working example.

As an aside, I maintain that some of these alignments are unnecessary --- the compiler should emit the appropriate directives --- but that's a separate discussion.